class was overriding in parent, that didn't even exist (anymore) was broken since Timer use adapters fixed

+++ source/ws/loewe/BattMan/Components/Timer/RunOnceTimer.php
@@ -4,16 +4,17 @@ namespace ws\loewe\BattMan\Components\Timer;
 
 use \ws\loewe\Woody\Components\Windows\AbstractWindow;
 use \ws\loewe\Woody\Components\Timer\Timer;
+use \ws\loewe\Woody\Event\TimeoutAdapter;
 
 class RunOnceTimer extends Timer {
-  
+
   public function __construct(\Closure $callback, AbstractWindow $window, $interval) {
     parent::__construct($callback, $window, $interval);
-  }
-  
-  public function run() {
-    parent::run();
-    
-    $this->destroy();
+
+    $this->addTimeoutListener(new TimeoutAdapter(function () {
+      if($this->counter == 1) {
+        $this->destroy();
+      }
+    }));
   }
 }
